geeting rid of models pkg
