+++ backend/controllers/account.mjs
@@ -1,4 +1,5 @@
 import AccountModel from "../model/account.mjs"; 
+import { } from "."
 
 class AccountController {
   getAccount(req, res) {
