--- backend/model/account.mjs
@@ -1,20 +1,0 @@
-import mongoose from 'mongoose';
-const { Schema, SchemaTypes} = mongoose;
-
-const accountSchema = new Schema({
-    id: SchemaTypes.String,
-    email: SchemaTypes.String,
-    password: SchemaTypes.String,
-    phoneNumber: {
-        type: SchemaTypes.String,
-        required: false,
-    },
-    smsEnabled: SchemaTypes.Boolean,
-    emailEnabled: SchemaTypes.Boolean,
-    createdAt: SchemaTypes.Number,
-    updatedAt: SchemaTypes.Number,
-});
-
-const AccountModel = mongoose.model("Account", accountSchema, "accounts");
-
-export default AccountModel;
